refactor point into int and double pair classes
